// src/app/models/user.ts
export class User {
        id: number;
        first_name: string;
        last_name: string;
        email: string;
        phone: string;
        organization: string;
        designation: string;
        salary: number;
        status: number;
        is_deleted: number;
        created_at: Date;
        updated_at: Date;
}